Move cnbi controller to its own dir
