+++ controllers/cnbi/customnbimage_controller_test.go
